Separat class into its own file, add require_relative code to spec file
